chat-app/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
distro==1.9.0
docutils==0.21.2
fastapi==0.115.8
fonttools==4.56.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==25.6.0
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.20
numpy==2.2.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
PyYAML==6.0.2
readme_renderer==44.0
redis==5.2.1
reflex==0.7.0
reflex-hosting-cli==0.1.35
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.15.2
SecretStorage==3.3.3
setuptools==75.8.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.22
starlette==0.45.3
starlette-admin==0.14.1
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
twine==6.1.0
typer==0.15.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
